Prototyping: adding categories to events to filter
timeline.js
Modified the compiled version of timeline.js to look at a global
variable timeline_categories to enable and disable categories to show.

example_json.json
Has every date modified to include a "category": [] attribute to specify an
array of categories to which it belongs.

example_json.html
Has been modified to demo this.

Currently, the data is reloaded everytime a category selection changes,
which is really slow. ideally filters should be applied to already loaded
json data.